pyproject.toml
[project]
name = "scripthut"
version = "0.1.0"
description = "Web interface to start and track jobs on remote systems (Slurm, ECS, AWS Batch)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "asyncssh>=2.14.0",
    "jinja2>=3.1.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
scripthut = "scripthut.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/scripthut"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["venv", ".venv"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
